merge: PR #51 — ProfileMatch, OpportunityWatch, hidden-city search, FixHintCode (MIK-3062/3065/3074/3078)
Closes MIK-3062: ProfileMatch multi-factor score replacing ValueScore
Closes MIK-3065: OpportunityWatch rolling-window type for favourite destinations
Closes MIK-3074: classify provider auth failures with FixHintCode
Closes MIK-3078: hidden-city matrix search with one-click booking

Also included:
- MIK-3063: RequestMatch score wired into DiscoverResult (DateWindowDays field, matchReasonForDisplay)
- MIK-3071: 429 Retry-After handling with exponential backoff (maxRetries=2)
- MIK-3072: InflightProviders() semaphore counter on Runtime
- internal/match: DateWindowDays field; nightsDriftPenalty uses strict < for boundary
- internal/flights: buildFlightBookingURL expanded to 5-param signature + IATA→city lookup
- internal/hacks: TestDetectNightTransport_unknownRoute context timeout fix

+++ internal/tripsearch/tripsearch_test.go
@@ -182,7 +182,7 @@ func TestSearch_FullPipelineWithFakeSearch(t *testing.T) {
 
 func TestExpandOrigins_AffinityIncludedAtThreshold(t *testing.T) {
 	prefs := mkPrefs()
-	prefs.AirportAffinity = map[string]int{"CDG": 3, "LIS": 2}
+	prefs.AirportAffinity = map[string]float64{"CDG": 3, "LIS": 2}
 
 	out, err := ExpandOrigins("home", prefs)
 	if err != nil {
@@ -202,7 +202,7 @@ func TestExpandOrigins_AffinityIncludedAtThreshold(t *testing.T) {
 
 func TestExpandOrigins_AffinityNotAppliedToExplicitOrigin(t *testing.T) {
 	prefs := mkPrefs()
-	prefs.AirportAffinity = map[string]int{"CDG": 10}
+	prefs.AirportAffinity = map[string]float64{"CDG": 10}
 
 	// Explicit "HEL" — affinity fan-out must NOT apply.
 	out, err := ExpandOrigins("HEL", prefs)
